Add golangci for basic linting with makefile automation.
As work starts to move this project towards tech-preview, having
linting available is helpful in identifying errors and code to
tidy. The rules have been copied from the Nomad project as a base
with some modifications based on codebase differences.

The makefile also now includes both a lint target, and a tools
target that can be used to install the lint tool used.