Enhance initial configuration abilities of autoscaler agent.
The autoscaler agent can now be figured in a number of ways in
order to provide a better UX and consistency with Nomad. This
change allows for the passing of multiple config files, a config
directory of files, or the use of CLI flags. When multiple files
are passed, they are merged in lexicographical order. If the user
also specifies CLI args, these will take precedence over values
passed by files.

The agent can now also make use of a default configuration, which
reduces the amount of boiler plate configuration needed.