agent: pass all Nomad config when dispensing plugins.
When performing SetConfig on plugins that call the Nomad API, a
small subset of potential options were being used. In order for
these plugins to work in environments which use ACLs and TLS,
all configuration from the agent should be passed down. The
plugins only accept generic map structures as configuration to
ensure a flexible interface. Therefore a number of helper funcs
have been added to perform translation between internal structs
and maps.