docs: update demo to use 0.0.1-techpreview1 docker build.
Now that the HashiCorp docker build has been pushed to docker hub
the demo can be updated to use this source when running. In
addition the job has had a couple of unsed config items removed
and also now sets up a service, so that the autoscaler health
endpoint can be used and seen from Consul as a health check.