plugins: allow target-value to handle zero count and target params
It is possible that the count of a task group we wish to scale is
0. It is also possible that the target the operator wishes the
metric maintain is zero. An example here would be a build system
queue, where the number of agents can scale to 0 if there is no
work to process. If work is added to the queue, the strategy is
to scale the build agents so that the queue metric is reduced to
the target of 0 items.