plugins: improve external plugin development UX and mimic Nomad.
The UX for developing external plugins was OK, but required some
knowledge of the go-plugin library in order to correctly setup
and serve the plugin. This change aims to lower the barrier to
entry on writing plugins by making the required code simpler and
also mimicking the Nomad plugin interface.

PluginInfo has been moved into the base package as part of this
work. This seems a better logic location for the definition as it
is the response object of a defined interface function inside this
package.