build: add makefile target to check go.mod status.
It is fairly easy for the go.mod file to become cluttered with
unrequired dependancies. This makefile addition therefore runs the
go mod command and checks if the go.mod file has changed as a
result to ensure this is caught during CI.

The target runs after the build phase as this also allows us to
catch dependancies that have not been added to the go.mod file
which are needed. This likely wont be hit during CircleCI runs,
but could be helpful for the development cycle.