policy: improve target block policy parsing and fix panics.
The target block of a scaling policy is opaque to Nomad and does
not go through any validation during job registration. It can
result in malformed but valid HCL being read by the autoscaler. It
is its responsibility to parse this carefully, avoiding panics
and returing sensible errors where possible.

This fix initially targeted a panic which occured if the name
param of a target block was blank where the block also included
a config section. When writing tests to cover this, a number of
other cases were uncovered and the scope of the change altered to
accomodate these.