prometheus: set config should error if address is not set.
When setting the configuration on the Prometheus APM plugin, it is
not required to pass the address config param. When the plugin
performs a query, it will error as the SDK does not add any
default address and so the HTTP call uses an address of "". This
change therefore suggests returning an error if the user specified
address is empty at the set config phase. This will provide
quicker user feedback on an error which will always occur
eventually.