plugin: add tests to strategy action.
Adds tests to the strategy Action set of functions. The tests
uncovered a bug and lint enhancement which are fixed within this
commit.

The lint enhancement changes the PushReason function to be
un-exported as it is currently only used within the strategy pkg.

The bug occured where a reason was added to the action in
situations where no reason previously existed. The meta reason
history was appended to from the empty reason string, before the
reason string was subsequently set.