policy: fix Nomad target parsing when target block is empty.
When parsing a scaling policy from Nomad which did not contain a
target block within the policy map, the outer target map which
contained both the job and group information was ignored. This
resulted in evaluation failures and rendered the policy incapable
of being used by the autoscaler.

The fix updates the target parsing to correctly handle the above
described situation. Whilst also figuring this out, it made sense
to remove the unnecessary additional target mapping within
canonicalizePolicy() as pointed out in comments by @lgfa29. This
requires a small change to the Nomad target plugin, but results in
a simpler workflow.