agent: support scaling cooldowns to aid oscillation migration.
Cooldown is now supported within a scaling policy as a time
duration. If the scaling policy omits the cooldown field, a
default value will be used. The default value can be controlled
by the CLI and has a default value of 5mins.

In the event the agent successfully registers a scaling action,
the policy manager will be instructed to place the policy eval
handler in cooldown. The cooldown is enforced by performing a
timed block using time.Timer.

Target plugins can optionally return a meta entry to detail the
last scaling event from their prespective. This allows the
autoscaler to enforce cooldowns, even when scaling actions have
been triggered manually by operators. The flow is similar to that
mentioned above, the only difference being that the block period
is based on the calculated remaning cooldown time from the last
event timestamp.