plugin: fix bug in ext. strategy plugins suggesting scale to zero.
The way in which stdlib encoding/gob works means a pointer to a
types default value will result in the RPC translation becoming
nil. This meant external strategy plugins were not able to suggest
scaling to zero, as the pointer to zero would be translated to nil
which inturn is classed as dry-run.

This change removes the use of a pointer for count within strategy
action struct. Allowing the use of zero to indicate scaling to
zero. Dry run situations use a special -1 count field, to indicate
a dry run workflow should be triggered.