plugin/target: allow operators to limit the amount of change.
Currently there is no method to limit the amount of change made in
a single evaluation. The only enforcements made are to ensure the
lower and upper limits are never breached. This change introduces
an optional target config parameter "max_change" which can limit
the maximum chnage in count actioned during a single evaluation.