plugins/target/aws-asg: Add new target plugin for AWS ASGs.
This commit adds to initial work for the new AWS AutoScaling Group
target plugin which allows for the scaling of AWS ASGs.

The plugin attempts to perform some basic validation that actions
taken against the AWS API are completed as desired. There is at
this point no reconciliation of failed steps, which should be
added at a later date.

The plugin is within the main repository as we do not currently
have a story of building and releasing plugins from separate repos
which should be looked at in the future. The plugin is also not
launched within the plugin manager, this will be added later when
all the cluster scaling work is tied together.