strategy: use single action for Run response rather than array.
The array response from a strategy run was legacy left over from
initial designs and has become redundant. This commit therefore
removes the use of an array in favour of a single action response.

The single response makes the agent evaluation code simpler and
will also make future multi-check work easier to implement.

It was discovered that the length of the response was used to
perform some evaluation logic. The strategy response object has
therefore been updated to include a direction. This is not a new
concept to the Autoscaler and was used in the target-value
strategy plugin originally. It is also used within infra provider
target plugins as they have to deal with calling APIs with
incremental count changes rather than absolute. I believe this is
therefore an acceptable change to the interface and provides wide
benefit.