agent: allow capturing of SIGHUP to reload policy sources
The file source in particular benefits greatly from being able
to perform reloads. This allows the source to re-read and decode
files which may have changed as well as process new or delete
scaling policies. In order to keep the behaviour as consistent as
possible across all sources the interface has been modified with
this new functionality. The Nomad source noops a reload, but in
the future we may want to use this to reload Nomad client config
params, in particular ACL tokens.