plugins: allow all plugins to inherit agents Nomad config.
Plugins such as the AWS ASG target plugin setup a Nomad client in
order to provide scaleutil help. If the operator is using a Nomad
cluster which requires non-default config options, it would be
preferable that they do not have to supply them twice, once to the
agent and again to the plugin.

This changes makes it so that plugin inherit the agents Nomad
configuration by default, reducing the overhead on operators. When
merging config options within the plugin config will take
preference. This allows the operator to provide a subset of params
which are considered specific to the plugin such as ACL token, but
inherit the generic options such as address.

The generic plugin config key "nomad_config_inherit" allows
operators to override default behaviour via a boolean value.