agent: move and fix default eval interval param to policy block. (#197)
The addition of a policy config block allows us to move and
evaluation interval parameter to this section. The param is only
used for policies and so makes sense in this section. A result of
this move fixes an incorrect detailing of the CLI flag which
referenced its old name.

Co-authored-by: Chris Baker <1675087+cgbaker@users.noreply.github.com>