scaleutils: protect Autoscaler node from scale in
Until the Nomad Autoscaler has the ability to reconcile actions
which have not reached a terminal state, the scaleutils should
protect the node the Nomad Autoscaler is running on from scalein
actions. This gives better assurance that the autoscaler wont
drain the node its on and then not be able to complete the scaling
action.

Even once the autoscaler has reconcile ability, there are some
interesting dicussions to be had about what logic to apply in the
situation where the autoscaler is running as a Nomad job and
identifies its own node for scaling in. In this case, we know it
will never be fully able to complete the action.