policy: validate key params of policy struct.
This adds a validate function to the Policy struct which allows
for the validation ensuring key params are not outside of the
expected.

Ideally both policy sources would use the same function, however,
the Nomad policy source performs validation on the Nomad API
policy struct so there is work to do if we wish to make the
change.