policy: fix issue when parsing non-default source name Nomad query.
When converting a Nomad APM short query to its long form, there
was an assumption that the plugin was named in its default manner
which caused problems where operators used their own naming
convention.

This change fixes that issue by passing the Nomad APM names to
the function which performs the query transformation. In addition
to simply fixing the issue, this change adds a new policy
processor that can be used to perform validation, canonicalization
and default setting work. I opted for this as it reduced the need
to duplicate work in the two sources. In this case I would have
had to add the APM name list to each source which seems silly and
continues to become annoying in the event we have other
functionality to add to policy parsing and such.