scaleutils: add post scale in task to optionally purge nodes.
When performing cluster scaling and removing nodes from the pool
of compute, an operator may optionally wish to have the terminated
node removed from the Nomad node list. This can be performed by
using the node purge API call and is now included as part of the
scaleutils scale in post tasks run.

The default behaviour is that the purge is not run. The operator
can change this by adding a config option "node_purge" to the
target config mapping. The RunPostScaleInTasks func handles
checking the target configuration, so that this work does not need
to be repeated across all cluster scaling plugins.