sdk: move common structs and consts into sdk pkg.
Type declarations have until now been placed within the package
where they are most typically used. This worked well during the
early phase of the project, however, as the codebase matures and
grows this become harder to manage and results in complex
dependency trees.

In order to make the codebase easier to work with both as an
internal and external developer, an sdk pkg has been created. This
pkg is designed to host much of the common declarations and should
not import from within the nomad-autoscaler project. The code moved
initially forms a good base, but is not itended to be everything
that may end up within the SDK pkg.