strategy: update interface to accept and respond with evaluation.
The strategy calculation within the autoscaler is one portion of
the policy evaluation pipeline. Internal discussion has concluded
that it is appropriate to pass the evaluation to the strategy and
have it populate the action.

The new ScalingEvaluation encompasses all required information
the autoscaler needs when performing an autoscaling evaluation of
a policy.