scaleutils: pass ignoreSystem drainspec param to drain monitor.
The IgnoreSystemJobs DrainSpec param was not being passed to the
drain monitor as this was previously hard-coded and not exposed
to users.

Recent updates had exposed this as a config param to users via
scaling policies and is sent correctly when calling the Nomad API
and draining a node. However, the drain monitor was still using
a hardcoded false value, meaning it would wait endlessly for all
allocs to be drained, including system allocs which were never
going to be drained.