plugins/target/aws-asg: correctly order region config setting.
A user specific region configuration option should take precedence
over other options. This change therefore updates the flow in
which the AWS config is populated. In situation where AWS and the
operator do not provide an option, the default is applied.

The change itself is hard to automatically test at the moment as
the behaviour changes depending whether you have the EC2 metadata
service available. The test added is aimed at situations where the
plugin is not running with the EC2 metadata service available.