helper: add metric pkg to allow setting default labels.
When running in HA and therefore when each agent has a unique ID,
it will be useful to have this ID as a default label. This allows
easier understanding across telemetry, logs and service discovery
processes.

The new package wraps the metrics library calls, ensuring default
and additional labels are always added where required.