scaleutils: fix least busy selector on clusters older than v1.0.0
This adds additional API calls when performing the least busy node
selector as a best effort to work on cluster that running Nomad
versions before v1.0.0.

v1.0.0 including API functionality which allows node resources to
be include in the stub list node object which is used when
performing least busy calculations. In version priror to this, the
information is only available within the node info object.