HA: add initial consistent hashring implementation.
The ring focusses on consistency and doesn't include functionality
such as load balancing.

The consistent hashring is used for partitioning policies between
autoscaler agents. It keeps an internal tracking of load
distribution across known members, allowing telemetry to identify
inconsistency.

The current implementation is specific to Consul and the manner in
which agents are discovered. If we ever moved to an alternative
discovery mechanism, we would need some additional functions but
the logic is sound.