add support for grouped checks
Grouped checks are checks that have the same value for the new `group`
configuration. Checks that are part of the same group have their results
compared to one another resulting in a single action representing all
checks. This final result is then compared to the rest of the checks.

Grouped checks should be used when multiple checks use the same or
correlated data as input. In these situations it's expected that only
one check (or a handful of them) will actually have enough data to make
a decistion.

For example, consider these two checks that monitore the percentage of
memory usage in a cluster:

```
check "memory-low" {
  source = "prometheus"
  query  = "..."

  strategy "threshold" {
    upper_bound = 30
    delta       = -1
  }
}

check "memory-high" {
  source = "prometheus"
  query  = "..."

  strategy "threshold" {
    lower_bound = 70
    delta       = 1
  }
}
```

If memory usage is at 20% the `memory-low` check should be enacted and
reduce the cluster size by 1. But sice `memory-high` doesn't have enough
data within its range, it will return a `none` scaling action,
preventing the cluster from scaling in since `none` is a safer option
than `down`.

By grouping these two checks together (since they act based on the same
data), the `none` result is ignored and the `down` action is returned,
allowing the cluster to scale in.

The only time `none` should be considered in a group is if all checks in
the group return `none`. In the example above, if the memory usage was
at 50% both checks would return `none` since the cluster is in a good
spot.

In this scenario, the safe option should be to do nothing. Returning a
`none` for the group prevents an accidental scale in if a check from
another group return `down`.

If all checks return a `nil` action, the group is ignored.