ci: fix nightly
The `github.event` context is not available to workflows dispatch calls,
so the `repo` env var was empty.

Using the `PKG_NAME` avoids this problem and also provides a more
consistent artifact naming, since all other are based on this value
already.

GitHub release pages also render the line break in the Markdown,
resulting in a weird narrow output.