improve 404 handling on jobs and policies
it's totally normal for jobs to go away,
along with their policies, so that should not be
considered an Error.

often when a policy is deleted, the blocking query
on /v1/scaling/policy/{policy-id} passes a 404
to the policy Handler, before the Manager can stop
it after receiving an update from /v1/policies

similar for job scaling events at /v1/job/{job-id}/scale
when a job goes away.