recover file policies after manager soft-restarts
Fixes "failed to read policy in time" errors, at
least some of which occur due to Handlers getting
re-started, but then not receiving file policies,
because they "already exist" from the point of view
of the file policy Source.

When the policy Manager creates and runs new Handlers,
the new Handler needs to receive a policy on its h.ch,
or h.handleTick() will have a nil currentPolicy, which
results in "timeout: failed to read policy in time"

So, each time Handler asks a Source to MonitorPolicy(),
the Source needs to send a policy back to the Handler
(via MonitorPolicyReq.ResultCh: h.ch), even if the
Source already knows the policy from a previous run.