threhsold: clarify the behaviour of bounds (#815)
The `lower_bound` value is always considered inclusive and the
`upper_bound` value is always exclusive. This helps prevent overlapping
or a dead zone when regions start or end right next to each other.

Also use `math.Inf` instead of `MaxFloat64` to allow a value that
matches `MaxFloat64` to be considered within bounds.