ha: add support for specifying the lock namespace
When HA mode is enabled, the leader lock must be managed in a specific
namespace, but when setting the `nomad.namespace` configuration to the
wildcard `*` value the Nomad API client was trying to write to this
namespace value, which is not allowed.

This commit adds a new `high_availability.lock_namespace` configuration
to allow operators to specify the exact lock variable namespace.