policy: add `query_window_offset` configuration (#850)
The Nomad Autoscaler uses the current time as the end of the query
window. This can cause instability, specially for strategies that only
take the most recent data point into account, because the latest data
may still be processing.

`query_window_offset` defines a time offset applied to the query window
to push it back in time. A short `query_window_offset` value can help
with the unstable readings, and a long value can be used for predictive
scaling based on historical data (for example, from `7d` ago).