Add scaling_min/scaling_max to checks
This allows for a check to be disabled if the current count of
allocations is not within its scaling_min;scaling_max bounds.
A disabled check will return ScaleNone. Ideally, we would completely
ignore it, in a way that it would be like it is not even defined in the
spec.