Fix race between worker and policy hander (#1071)
Because of bug/feature described here #1070 we had a lot of scaling and cooldown actions. With thousands of policies, we've noticed that sometimes all broker workers get stuck.

After some debugging, I believe this is a result of a race between broker worker goroutine and policy handler goroutine.

Here's what's happening:
After each scaling action, the worker enforces a cooldown on the policy handler.
Policy handler gets this information via the cooldownCh channel here.
However, if the policy handler is already executing handleTick method, it can notice the scale event and also start cooldown here. This will block the whole Run method, and reading from cooldownCh will be blocked for a configured cooldown time.
After cooldown time, handleTick will finish, and will read from cooldownCh and will unlock the worker and block again.

So there are two issues here:

Broker worker is blocked. We set cooldown to 5min so it's noticeable.
The Policy handler is on cooldown twice. Once because of handleTick and once because of cooldownCh
This PR fixes only the first issue. By making cooldownCh channel buffered, we avoid blocking the worker.