recover from validation errors in policy monitoring (#1125)
* recover from validation errors in policy monitoring

The `monitorPolicies` method of the manager can receive Nomad API scaling
policies with a nil inner policy (for example, if the policy is intended not to
be used for autoscaling). In this case validation of the policy fails but the
monitor is continuing as though it hadn't because it's not continuing to the
next ID when hitting a recoverable error.

Fixes: https://github.com/hashicorp/nomad-autoscaler/issues/1124

* func: add testing for errors to avoid unexpected panics

* func: add a more descriptive message for errors

* fix: set the correct error message

---------

Co-authored-by: Juanadelacuesta <8647634+Juanadelacuesta@users.noreply.github.com>