policy: only count should be used to determine if scaling is required (#1221)
In d2df3081464b we changed the calculation for determining if scaling is needed
to only include the count if direction was `sdk.ScaleDirectionNone` in order to
support Dynamic Application Sizing (DAS). But in practice we can get a up/down
direction with a count that's unchanged from non-DAS policies in cases where
we've reached the limit. This causes deployment loops as we keep re-submitting
the scaling request.

Ref: https://github.com/hashicorp/nomad-autoscaler/commit/d2df3081464bcd5eee10d9991ca3a0cd456e2899

Co-authored-by: Waquid Valiya Peedikakkal <waquidvp@gmail.com>