build: don't lint when building dev target (#1222)
Using `golangci-lint` is quite slow and memory hungry, so using it for every
dev build slows down bug fixes. We run it in CI and developers should have
linting wired up in their editors already. Doing the `hclfmt` is quick and
cheap, and we do it on the core product already.

Also, there's no need to reinstall all the tools every time we run the
linter. They'll already be installed.