policy: handle recommendations independently of scaling no-op
In #1221 we changed the check for whether scaling was needed to only rely on the
change in count and not the direction, in order to account for non-DAS policies
that have hit their limit and have a direction but no change in count. But this
failed to account for plugins that return an essentially zero-value
`sdk.ScalingAction` where the direction is none. In this case, the logic we
added in #1221 will cause the autoscaler to scale down to the minimum value
instead of keep it steady.

Because DAS sets `sdk.ScaleDirectionNone` there's no reasonable way to tell the
difference between a plugin that's left the count unset (which then gets capped
to the minimum) and a plugin like DAS that's setting a specific count. Fix this
by adding a new `sdk.ScaleDirectionRecommendation` value to the direction enum,
which will be used by DAS built in plugins.

Ref: https://github.com/hashicorp/nomad-autoscaler/pull/1221
Fixes: https://github.com/hashicorp/nomad-autoscaler/issues/1211